Refactor BundleSet out of SolsticeManifest.
