+++ solstice/src/main/java/dev/equo/solstice/SolsticeManifest.java
@@ -24,15 +24,11 @@ import java.util.Enumeration;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.Set;
-import java.util.TreeMap;
-import java.util.function.Function;
 import java.util.jar.Manifest;
 import javax.annotation.Nullable;
 import org.eclipse.osgi.util.ManifestElement;
 import org.osgi.framework.BundleException;
 import org.osgi.framework.Constants;
-import org.slf4j.Logger;
 
 /**
  * Parses a jar manifest, removing some fine-grained details for the purpose of simplifying the
@@ -44,170 +40,6 @@ import org.slf4j.Logger;
  * </ul>
  */
 public class SolsticeManifest {
-	/** Represents a closed universe of OSGi bundles. */
-	public static class BundleSet {
-		private final List<SolsticeManifest> bundles;
-
-		private BundleSet(List<SolsticeManifest> bundles) {
-			this.bundles = bundles;
-		}
-
-		public List<SolsticeManifest> getBundles() {
-			return Collections.unmodifiableList(bundles);
-		}
-
-		public Map<String, List<SolsticeManifest>> bySymbolicName() {
-			return groupBundles(manifest -> Collections.singletonList(manifest.getSymbolicName()));
-		}
-
-		public Map<String, List<SolsticeManifest>> byExportedPackage() {
-			Map<String, List<SolsticeManifest>> manifests = groupBundles(SolsticeManifest::getPkgExports);
-			// fragments export the same packages as their parent, no need to report those conflicts
-			manifests.replaceAll(
-					(pkg, bundles) -> {
-						if (bundles.size() > 1) {
-							long numNonFragments =
-									bundles.stream().filter(SolsticeManifest::isNotFragment).count();
-							if (numNonFragments == 1) {
-								return Collections.singletonList(
-										bundles.stream().filter(SolsticeManifest::isNotFragment).findFirst().get());
-							}
-						}
-						return bundles;
-					});
-			return manifests;
-		}
-
-		public Map<String, List<SolsticeManifest>> calculateMissingBundles(Set<String> available) {
-			return calculateMissing(available, SolsticeManifest::getRequiredBundles);
-		}
-
-		public Map<String, List<SolsticeManifest>> calculateMissingPackages(Set<String> available) {
-			return calculateMissing(available, SolsticeManifest::getPkgImports);
-		}
-
-		private Map<String, List<SolsticeManifest>> calculateMissing(
-				Set<String> available, Function<SolsticeManifest, List<String>> neededFunc) {
-			TreeMap<String, List<SolsticeManifest>> map = new TreeMap<>();
-			for (SolsticeManifest bundle : bundles) {
-				for (var needed : neededFunc.apply(bundle)) {
-					if (!available.contains(needed)) {
-						mapAdd(map, needed, bundle);
-					}
-				}
-			}
-			return map;
-		}
-
-		private Map<String, List<SolsticeManifest>> groupBundles(
-				Function<SolsticeManifest, List<String>> groupBy) {
-			TreeMap<String, List<SolsticeManifest>> map = new TreeMap<>();
-			for (SolsticeManifest bundle : bundles) {
-				for (String extracted : groupBy.apply(bundle)) {
-					mapAdd(map, extracted, bundle);
-				}
-			}
-			return map;
-		}
-
-		private static <K> void mapAdd(
-				Map<K, List<SolsticeManifest>> map, K key, SolsticeManifest value) {
-			var oldValue = map.putIfAbsent(key, Collections.singletonList(value));
-			if (oldValue != null) {
-				if (oldValue.size() == 1) {
-					var newList = new ArrayList<SolsticeManifest>(2);
-					newList.addAll(oldValue);
-					oldValue = newList;
-					map.put(key, newList);
-				}
-				oldValue.add(value);
-			}
-		}
-
-		/**
-		 * Sends warnings to logger, then modifies every manifest to resolve all these warnings.
-		 *
-		 * <ul>
-		 *   <li>multiple bundles with same symbolic name (resolved by first one on the classpath, since
-		 *       that's what <code>Class.forName</code> will do)
-		 *   <li>multiple bundles which export the same package (resolved by first one on the classpath,
-		 *       since that's what <code>Class.forName</code> will do)
-		 *   <li>required bundle which is not present (resolved by removing requirement from the
-		 *       SolsticeManifest)
-		 *   <li>imported package which is not present (resolved by removing import from the
-		 *       SolsticeManifest)
-		 * </ul>
-		 */
-		public void warnAndModifyManifestsToFix(Logger logger) {
-			var bySymbolicName = bySymbolicName();
-			var byExportedPackage = byExportedPackage();
-			bySymbolicName.forEach(
-					(symbolicName, manifests) -> {
-						if (manifests.size() > 1) {
-							logger.warn("Multiple bundles with same symbolic name: " + symbolicName);
-							for (SolsticeManifest manifest : manifests) {
-								logger.warn("  - " + manifest.jarUrl);
-							}
-						}
-					});
-			byExportedPackage.forEach(
-					(pkg, manifests) -> {
-						if (manifests.size() > 1) {
-							int numSplitPkgs = 0;
-							for (SolsticeManifest manifest : manifests) {
-								if (pkg.startsWith("META-INF.versions.")) {
-									// just an artifact of multijar
-									return;
-								}
-								var element =
-										manifest.pkgExportsRaw().stream()
-												.filter(e -> e.getValue().equals(pkg))
-												.findFirst()
-												.get();
-								String mandatory = element.getDirective("mandatory");
-								if (mandatory != null) {
-									if ("split".equals(element.getAttribute(mandatory))) {
-										++numSplitPkgs;
-									}
-								}
-								String uses = element.getDirective("uses");
-								if (uses != null) {
-									++numSplitPkgs;
-								}
-							}
-							if (numSplitPkgs < manifests.size() - 1) {
-								logger.warn("Multiple bundles exporting the same package: " + pkg);
-								for (SolsticeManifest manifest : manifests) {
-									logger.warn("  - " + manifest.jarUrl);
-								}
-							}
-						}
-					});
-			var missingBundles = calculateMissingBundles(bySymbolicName.keySet());
-			missingBundles.forEach(
-					(missing, neededBy) -> {
-						logger.warn("Missing required bundle " + missing + " needed by " + neededBy);
-					});
-			var missingPackages = calculateMissingPackages(byExportedPackage.keySet());
-			missingPackages.forEach(
-					(missing, neededBy) -> {
-						if (missing.equals("kotlin") || missing.startsWith("kotlin.")) {
-							return;
-						}
-						if (missing.startsWith("javax.") || missing.startsWith("java.")) {
-							return;
-						}
-						if (missing.startsWith("org.xml.") || missing.startsWith("org.w3c.")) {
-							return;
-						}
-						logger.warn("Missing imported package " + missing + " imported by " + neededBy);
-					});
-			for (var bundle : bundles) {
-				bundle.requiredBundles.removeAll(missingBundles.keySet());
-				bundle.pkgImports.removeAll(missingPackages.keySet());
-			}
-		}
-	}
 
 	/** Finds every OSGi bundle on the classpath and returns it inside a {@link BundleSet}. */
 	public static BundleSet discoverBundles() {
@@ -232,8 +64,8 @@ public class SolsticeManifest {
 	private final int classpathOrder;
 	private final @Nullable String symbolicName;
 	private final LinkedHashMap<String, String> headersOriginal = new LinkedHashMap<>();
-	private final List<String> requiredBundles;
-	private final List<String> pkgImports;
+	final List<String> requiredBundles;
+	final List<String> pkgImports;
 	private final List<String> pkgExports;
 	private final boolean lazy;
 
@@ -282,7 +114,7 @@ public class SolsticeManifest {
 
 	private List<ManifestElement> pkgExportsRaw;
 
-	private List<ManifestElement> pkgExportsRaw() {
+	List<ManifestElement> pkgExportsRaw() {
 		if (pkgExportsRaw == null) {
 			pkgExportsRaw =
 					Arrays.asList(
@@ -295,7 +127,7 @@ public class SolsticeManifest {
 		return pkgExportsRaw;
 	}
 
-	private boolean isNotFragment() {
+	boolean isNotFragment() {
 		return !headersOriginal.containsKey(Constants.FRAGMENT_HOST);
 	}
